+++ packages/frontend/src/router/index.js
@@ -269,6 +269,15 @@ const routes = [
       title: 'Not Found | Speckle'
     },
     component: () => import('@/cleanup/pages/common/NotFound.vue')
+  },
+  // NOTE: only used for local debugging of the full embed app.
+  // (I might be missing something obvious though!)
+  {
+    path: '/embed*',
+    meta: {
+      title: 'Embed View | Speckle'
+    },
+    component: () => import('@/embed/EmbedViewer.vue')
   }
 ]
